Allow raw query strings to be added to requests
In some scenarios[1], it may be valuable to be able to pass
raw/pre-encoded query string parameters with a request. A recent change
to superagent forced all qs params to be utf8 encoded, which changed the
previous behavior of superagent. This commit adds another function,
`queryRaw`, to allow adding query string parameters that are NOT
re-encoded in any way prior to being passed as part of the request.

I thought it best to have an explicit approach to doing this to make
sure the developer intended this behavior, rather than trying to deduce
which querystring values should be encoded via other means.

[1] One example: when testing a server that uses a non-UTF8 URL encoding

See discussion here for more context:

https://github.com/visionmedia/superagent/commit/629bf9b10d26e2e6b44aea11b197ef8fea99a9c5