Fix IE 10/11 issue with `XMLHttpRequest.send(undefined)`
Fixes an issue where `Request#end` would call `XMLHttpRequest.send(undefined)`, causing an issue in Internet Explorer 10 and 11. In these browsers, `XMLHttpRequest.send(undefined)` will cause a request body of "undefined" to be sent by the browser (e.g. on a POST). This causes the `XMLHttpRequest` to assume a request body of type `text/plain` (as Content-Type), which confuses some server software.

Before this change, calling `Request#end` without supplying a request body would cause a call to `XMLHttpRequest.send(undefined)`. After the change, `Request#end` calls `XMLHttpRequest.send(data)` if data is defined for the request, or `XMLHttpRequest.send()` if no data is defined.

This change is in compliance with the "Level 1" and "Living Standard" standards documents for XMLHttpRequest (https://xhr.spec.whatwg.org/#the-send()-method), which permits a call to `send()` without parameters. It also appears that calling `send(undefined)` is not explicitly permitted by the standard, although most browsers seem to treat `null` and `undefined` equally.

I've added a test to verify the behavior. Before the fix, I verified that the test did not run on these browsers prior to the fix (using the Sauce Labs setup locally).